% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StackedDensityPlot.R
\name{densityCal}
\alias{densityCal}
\title{Internal density calculation function serves for \code{stackDenistyPlot}}
\usage{
densityCal(data, kernel = c("gaussian", "epanechnikov", "rectangular",
  "triangular", "biweight", "cosine", "optcosine"), bw = "nrd0", adjust = 1,
  reomoveOutliers = FALSE)
}
\arguments{
\item{data}{Input data frame including variable columns and stackFactor column.}

\item{kernel}{Kernel applied for density calculation, includes "gaussian", "epanechnikov", "rectangular", "triangular", "biweight", "cosine", "optcosine".}

\item{bw}{The smoothing bandwidth to be used. The default, "nrd0", has remained the default for historical and compatibility reasons, rather than as a general recommendation, where e.g., "SJ" would rather fit.}

\item{adjust}{The bandwidth used is actually adjust*bw.}

\item{reomoveOutliers}{If coerce the outliers to 1.5*IQR level.}
}
\value{
data frame containing stackName, x, y and densityName
}
\description{
Internal density calculation function serves for \code{stackDenistyPlot}
}
\examples{
data <- iris
data$stackFactor <- data$Species
data$Species <- NULL
dr <- densityCal(data = data)
}
