% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StackedDensityPlot.R
\name{stackedDenistyPlot}
\alias{stackedDenistyPlot}
\title{A function to create stacked density plot using ggplot}
\usage{
stackedDenistyPlot(data, densityColNames, stackFactor, stackColor,
  kernel = c("gaussian", "epanechnikov", "rectangular", "triangular",
  "biweight", "cosine", "optcosine"), bw = "nrd0", adjust = 1,
  reomoveOutliers = FALSE, stackRotation = 0, stackSeperation = "auto",
  x_text_size = 2, strip_text_size = 7, legend_text_size = 0.5,
  legendRow = 1, legend_title = "stackName", legend_title_size = 1,
  legend_key_size = 0.2)
}
\arguments{
\item{data}{Data frame containing columns for density calculation.}

\item{densityColNames}{The column names of those passed to density calculation.}

\item{stackFactor}{A factor to create the stacks, length equalt to number of rows in data.}

\item{stackColor}{Default use rainbow colour palette, but can be specified manually.}

\item{kernel}{kernel applied for density calculation, includes "gaussian", "epanechnikov", "rectangular", "triangular", "biweight", "cosine", "optcosine".}

\item{bw}{The smoothing bandwidth to be used. The default, "nrd0", has remained the default for historical and compatibility reasons, rather than as a general recommendation, where e.g., "SJ" would rather fit.}

\item{adjust}{The bandwidth used is actually adjust*bw.}

\item{reomoveOutliers}{If coerce the outliers to 1.5*IQR level.}

\item{stackRotation}{How many degrees to rotate the density stacks.}

\item{stackSeperation}{Adjust the stack height intervels, use "auto" for automatic adjustment; or specify a value form 0 to 1 times the median height of to all density plots.}

\item{x_text_size}{Adjust the text size on x ticks.}

\item{strip_text_size}{Adjust the text size on strip.}

\item{legend_text_size}{Adjust the text size on legend.}

\item{legendRow}{The number of rows of the legend.}

\item{legend_title}{The title of the legend.}

\item{legend_title_size}{The size of the legend title.}

\item{legend_key_size}{Adjust the size of the legend key.}
}
\value{
a ggplot object
}
\description{
A function to create stacked density plot using ggplot
}
\examples{
data <- iris[ ,1:4]
densityColNames <- colnames(data)
stackFactor <- iris[ ,5]
stackedDenistyPlot(data, densityColNames, stackFactor, strip_text_size = 9, legend_text_size = 0.8, legend_key_size = 0.15)
}
